Update deletion of student without mark

+++ student_grading/app/Http/Controllers/StudentController.php
@@ -105,7 +105,9 @@ class StudentController extends Controller
             abort(403);
         }
         $mark = Mark::where('student_id',  $student->id)->first();
-        $mark->delete();
+        if($mark){
+            $mark->delete();
+        }
         $student->delete();
 
         return to_route('student.index')->with('message', 'Student Deleted');
